Keyspace should be set from config
set_keyspace() when called without any arguments should set the
original keyspace set in the config. Instead it was being set to
whatever was set as self._keyspace_.

This is a problem if set_keyspace() was called earlier with some
argument. In this case self._keyspace_ would be set to a new
keyspace and calling the method without arguments will not set
the original keyspace again.